move token manager to services-core since it will be consumed by lambdas package and it needs to be isomorphic
