+++ server/routerlicious/packages/routerlicious-base/src/alfred/routes/api/index.ts
@@ -12,12 +12,12 @@ import {
 	IProducer,
 	ITenantManager,
 	IThrottler,
+	IJsonWebTokenManager,
 } from "@fluidframework/server-services-core";
 import cors from "cors";
 import { Router } from "express";
 import { Provider } from "nconf";
 import { IAlfredTenant } from "@fluidframework/server-services-client";
-import { IJsonWebTokenManager } from "@fluidframework/server-services-utils";
 import * as api from "./api";
 import * as deltas from "./deltas";
 import * as documents from "./documents";
