+++ server/routerlicious/packages/routerlicious-base/src/alfred/routes/index.ts
@@ -12,11 +12,11 @@ import {
 	ICache,
 	ICollection,
 	IDocument,
+	IJsonWebTokenManager,
 } from "@fluidframework/server-services-core";
 import { Router } from "express";
 import { Provider } from "nconf";
 import { IAlfredTenant } from "@fluidframework/server-services-client";
-import { IJsonWebTokenManager } from "@fluidframework/server-services-utils";
 import * as api from "./api";
 
 export interface IRoutes {
