+++ server/routerlicious/packages/routerlicious-base/src/alfred/runner.ts
@@ -19,11 +19,12 @@ import {
 	IThrottleAndUsageStorageManager,
 	IWebServer,
 	IWebServerFactory,
+	IJsonWebTokenManager,
+	IWebSocketTracker,
 } from "@fluidframework/server-services-core";
 import { Provider } from "nconf";
 import * as winston from "winston";
 import { createMetricClient } from "@fluidframework/server-services";
-import { IJsonWebTokenManager, IWebSocketTracker } from "@fluidframework/server-services-utils";
 import { IAlfredTenant } from "@fluidframework/server-services-client";
 import { Lumberjack } from "@fluidframework/server-services-telemetry";
 import { configureWebSocketServices } from "@fluidframework/server-lambdas";
