+++ server/routerlicious/packages/routerlicious-base/src/alfred/runnerFactory.ts
@@ -103,8 +103,8 @@ export class AlfredResources implements core.IResources {
 		public throttleAndUsageStorageManager?: core.IThrottleAndUsageStorageManager,
 		public verifyMaxMessageSize?: boolean,
 		public redisCache?: core.ICache,
-		public socketTracker?: utils.IWebSocketTracker,
-		public tokenManager?: utils.IJsonWebTokenManager,
+		public socketTracker?: core.IWebSocketTracker,
+		public tokenManager?: core.IJsonWebTokenManager,
 	) {
 		const socketIoAdapterConfig = config.get("alfred:socketIoAdapter");
 		const httpServerConfig: services.IHttpServerConfig = config.get("system:httpServer");
@@ -397,8 +397,8 @@ export class AlfredResourcesFactory implements core.IResourcesFactory<AlfredReso
 
 		// Set up token revocation if enabled
 		const tokenRevocationEnabled: boolean = config.get("tokenRevocation:enable") as boolean;
-		let socketTracker: utils.IWebSocketTracker | undefined;
-		let tokenManager: utils.IJsonWebTokenManager | undefined;
+		let socketTracker: core.IWebSocketTracker | undefined;
+		let tokenManager: core.IJsonWebTokenManager | undefined;
 		if (tokenRevocationEnabled) {
 			socketTracker = new utils.WebSocketTracker();
 			tokenManager = new utils.DummyTokenManager();
