+++ server/routerlicious/packages/services-core/src/index.ts
@@ -142,3 +142,8 @@ export {
 export { TokenGenerator } from "./token";
 export { clientConnectivityStorageId, IUsageData, signalUsageStorageId } from "./usageData";
 export { IZookeeperClient, ZookeeperClientConstructor } from "./zookeeper";
+export {
+	IWebSocketTracker,
+	IJsonWebTokenManager,
+	createCompositeTokenId,
+} from "./tokenManager"
