+++ server/routerlicious/packages/services-utils/src/auth.ts
@@ -15,11 +15,10 @@ import {
 	validateTokenClaimsExpiration,
 	canRevokeToken,
 } from "@fluidframework/server-services-client";
-import type { ICache, ITenantManager } from "@fluidframework/server-services-core";
+import type { ICache, ITenantManager, IJsonWebTokenManager } from "@fluidframework/server-services-core";
 import type { RequestHandler, Request, Response } from "express";
 import type { Provider } from "nconf";
 import { getLumberBaseProperties, Lumberjack } from "@fluidframework/server-services-telemetry";
-import { IJsonWebTokenManager } from "./tokenManager";
 
 /**
  * Validates a JWT token to authorize routerlicious.
