+++ server/routerlicious/packages/services-utils/src/index.ts
@@ -38,9 +38,6 @@ export {
 export { IThrottleMiddlewareOptions, throttle } from "./throttlerMiddleware";
 export { WinstonLumberjackEngine } from "./winstonLumberjackEngine";
 export {
-	IWebSocketTracker,
-	IJsonWebTokenManager,
 	WebSocketTracker,
-	createCompositeTokenId,
 	DummyTokenManager,
 } from "./tokenManager";
